WIP simpler useTracker
